Update pro package
