+++ docs/data/chartsApiPages.ts
@@ -200,11 +200,6 @@ const chartsApiPages: MuiPage[] = [
     title: 'PiePlot',
   },
   {
-    pathname: '/x/api/charts/responsive-chart-container-pro',
-    title: 'ResponsiveChartContainerPro',
-    plan: 'pro',
-  },
-  {
     pathname: '/x/api/charts/scatter',
     title: 'Scatter',
   },
