--- docs/pages/x/api/charts/responsive-chart-container-pro.json
@@ -1,30 +1,0 @@
-{
-  "props": {
-    "onZoomChange": {
-      "type": { "name": "func" },
-      "signature": {
-        "type": "function(zoomData: Array<ZoomData>) => void",
-        "describedArgs": ["zoomData"]
-      }
-    },
-    "zoom": {
-      "type": {
-        "name": "arrayOf",
-        "description": "Array&lt;{ axisId: number<br>&#124;&nbsp;string, end: number, start: number }&gt;"
-      }
-    }
-  },
-  "name": "ResponsiveChartContainerPro",
-  "imports": [
-    "import { ResponsiveChartContainerPro } from '@mui/x-charts-pro/ResponsiveChartContainerPro';",
-    "import { ResponsiveChartContainerPro } from '@mui/x-charts-pro';"
-  ],
-  "classes": [],
-  "spread": true,
-  "themeDefaultProps": null,
-  "muiName": "MuiResponsiveChartContainerPro",
-  "filename": "/packages/x-charts-pro/src/ResponsiveChartContainerPro/ResponsiveChartContainerPro.tsx",
-  "inheritance": null,
-  "demos": "<ul><li><a href=\"/x/react-charts/composition/\">Chart composition</a></li></ul>",
-  "cssComponent": false
-}
