--- docs/translations/api-docs/charts/responsive-chart-container-pro/responsive-chart-container-pro.json
@@ -1,11 +1,0 @@
-{
-  "componentDescription": "",
-  "propDescriptions": {
-    "onZoomChange": {
-      "description": "Callback fired when the zoom has changed.",
-      "typeDescriptions": { "zoomData": "Updated zoom data." }
-    },
-    "zoom": { "description": "The list of zoom data related to each axis." }
-  },
-  "classDescriptions": {}
-}
