+++ packages/x-charts-pro/src/ChartContainerPro/ChartContainerPro.test.tsx
@@ -3,9 +3,9 @@ import { expect } from 'chai';
 import { createRenderer, screen } from '@mui/internal-test-utils';
 import { LicenseInfo } from '@mui/x-license';
 import { sharedLicenseStatuses } from '@mui/x-license/useLicenseVerifier/useLicenseVerifier';
-import { ResponsiveChartContainerPro } from './ResponsiveChartContainerPro';
+import { ChartContainerPro } from './ChartContainerPro';
 
-describe('<ResponsiveChartContainerPro /> - License', () => {
+describe('<ChartContainerPro /> - License', () => {
   const { render } = createRenderer();
 
   beforeEach(() => {
@@ -17,9 +17,9 @@ describe('<ResponsiveChartContainerPro /> - License', () => {
   it('should render watermark when the license is missing', async () => {
     LicenseInfo.setLicenseKey('');
 
-    expect(() =>
-      render(<ResponsiveChartContainerPro series={[]} width={100} height={100} />),
-    ).toErrorDev(['MUI X: Missing license key.']);
+    expect(() => render(<ChartContainerPro series={[]} width={100} height={100} />)).toErrorDev([
+      'MUI X: Missing license key.',
+    ]);
 
     expect(await screen.findAllByText('MUI X Missing license key')).not.to.equal(null);
   });
