+++ packages/x-charts-pro/src/ChartContainerPro/ChartContainerPro.tsx
@@ -1,69 +1,39 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import { Watermark } from '@mui/x-license/Watermark';
 import { ChartContainerProps } from '@mui/x-charts/ChartContainer';
-import { ChartsSurface } from '@mui/x-charts/ChartsSurface';
-import { HighlightedProvider, ZAxisContextProvider } from '@mui/x-charts/context';
-import {
-  ChartsAxesGradients,
-  DrawingProvider,
-  InteractionProvider,
-  PluginProvider,
-  SeriesProvider,
-  AnimationProvider,
-} from '@mui/x-charts/internals';
-import { useLicenseVerifier } from '@mui/x-license/useLicenseVerifier';
+import { ResizableContainer } from '@mui/x-charts/internals';
 import { getReleaseInfo } from '../internals/utils/releaseInfo';
-import { CartesianProviderPro } from '../context/CartesianProviderPro';
-import { ZoomProps, ZoomProvider } from '../context/ZoomProvider';
+import { ChartDataProviderPro } from '../context/ChartDataProviderPro';
+import { ZoomProps } from '../context/ZoomProvider';
 import { useChartContainerProProps } from './useChartContainerProProps';
-
-const releaseInfo = getReleaseInfo();
 
 export interface ChartContainerProProps extends ChartContainerProps, ZoomProps {}
 
-const ChartContainerPro = React.forwardRef(function ChartContainer(
+const releaseInfo = getReleaseInfo();
+
+const ChartContainerPro = React.forwardRef(function ChartContainerPro(
   props: ChartContainerProProps,
   ref,
 ) {
-  const {
-    zoomProviderProps,
-    drawingProviderProps,
-    seriesProviderProps,
-    zAxisContextProps,
-    highlightedProviderProps,
-    cartesianProviderProps,
-    chartsSurfaceProps,
-    pluginProviderProps,
-    animationProviderProps,
-    children,
-  } = useChartContainerProProps(props, ref);
+  const { chartDataProviderProProps, resizableChartContainerProps, hasIntrinsicSize } =
+    useChartContainerProProps(props, ref);
 
-  useLicenseVerifier('x-charts-pro', releaseInfo);
+  if (props.height && props.width) {
+    return (
+      <React.Fragment>
+        <ChartDataProviderPro {...chartDataProviderProProps} />
+        <Watermark packageName="x-charts-pro" releaseInfo={releaseInfo} />
+      </React.Fragment>
+    );
+  }
 
   return (
-    <DrawingProvider {...drawingProviderProps}>
-      <AnimationProvider {...animationProviderProps}>
-        <PluginProvider {...pluginProviderProps}>
-          <ZoomProvider {...zoomProviderProps}>
-            <SeriesProvider {...seriesProviderProps}>
-              <CartesianProviderPro {...cartesianProviderProps}>
-                <ZAxisContextProvider {...zAxisContextProps}>
-                  <InteractionProvider>
-                    <HighlightedProvider {...highlightedProviderProps}>
-                      <ChartsSurface {...chartsSurfaceProps}>
-                        <ChartsAxesGradients />
-                        {children}
-                      </ChartsSurface>
-                    </HighlightedProvider>
-                  </InteractionProvider>
-                </ZAxisContextProvider>
-              </CartesianProviderPro>
-            </SeriesProvider>
-          </ZoomProvider>
-        </PluginProvider>
-      </AnimationProvider>
-    </DrawingProvider>
+    <ResizableContainer {...resizableChartContainerProps}>
+      {hasIntrinsicSize ? <ChartDataProviderPro {...chartDataProviderProProps} /> : null}
+      <Watermark packageName="x-charts-pro" releaseInfo={releaseInfo} />
+    </ResizableContainer>
   );
 });
 
@@ -91,9 +61,9 @@ ChartContainerPro.propTypes = {
    */
   disableAxisListener: PropTypes.bool,
   /**
-   * The height of the chart in px.
+   * The height of the chart in px. If not defined, it takes the height of the parent element.
    */
-  height: PropTypes.number.isRequired,
+  height: PropTypes.number,
   /**
    * The item currently highlighted. Turns highlighting into a controlled prop.
    */
@@ -131,6 +101,16 @@ ChartContainerPro.propTypes = {
    */
   plugins: PropTypes.arrayOf(PropTypes.object),
   /**
+   * The chart will try to wait for the parent container to resolve its size
+   * before it renders for the first time.
+   *
+   * This can be useful in some scenarios where the chart appear to grow after
+   * the first render, like when used inside a grid.
+   *
+   * @default false
+   */
+  resolveSizeBeforeRender: PropTypes.bool,
+  /**
    * The array of series to display.
    * Each type of series has its own specificity.
    * Please refer to the appropriate docs page to learn more about it.
@@ -154,9 +134,9 @@ ChartContainerPro.propTypes = {
     y: PropTypes.number,
   }),
   /**
-   * The width of the chart in px.
+   * The width of the chart in px. If not defined, it takes the width of the parent element.
    */
-  width: PropTypes.number.isRequired,
+  width: PropTypes.number,
   /**
    * The configuration of the x-axes.
    * If not provided, a default axis config is used.
