+++ packages/x-charts-pro/src/ChartContainerPro/useChartContainerProProps.ts
@@ -1,6 +1,6 @@
 'use client';
 import { useChartContainerProps } from '@mui/x-charts/internals';
-import { ZoomProviderProps } from '../context/ZoomProvider';
+import type { ChartDataProviderProProps } from '../context/ChartDataProviderPro';
 import type { ChartContainerProProps } from './ChartContainerPro';
 
 export const useChartContainerProProps = (
@@ -9,37 +9,20 @@ export const useChartContainerProProps = (
 ) => {
   const { zoom, onZoomChange, ...baseProps } = props;
 
-  const {
-    children,
-    drawingProviderProps,
-    seriesProviderProps,
-    cartesianProviderProps,
-    zAxisContextProps,
-    highlightedProviderProps,
-    chartsSurfaceProps,
-    pluginProviderProps,
-    animationProviderProps,
-    xAxis,
-    yAxis,
-  } = useChartContainerProps(baseProps, ref);
-
-  const zoomProviderProps: Omit<ZoomProviderProps, 'children'> = {
+  const chartDataProviderProProps: Pick<ChartDataProviderProProps, 'zoom' | 'onZoomChange'> = {
     zoom,
     onZoomChange,
-    xAxis,
-    yAxis,
   };
 
+  const { chartDataProviderProps, resizableChartContainerProps, hasIntrinsicSize } =
+    useChartContainerProps(baseProps, ref);
+
   return {
-    zoomProviderProps,
-    children,
-    drawingProviderProps,
-    pluginProviderProps,
-    seriesProviderProps,
-    cartesianProviderProps,
-    zAxisContextProps,
-    highlightedProviderProps,
-    chartsSurfaceProps,
-    animationProviderProps,
+    chartDataProviderProProps: {
+      ...chartDataProviderProps,
+      ...chartDataProviderProProps,
+    },
+    resizableChartContainerProps,
+    hasIntrinsicSize,
   };
 };
