--- packages/x-charts-pro/src/ResponsiveChartContainerPro/index.ts
@@ -1,1 +1,0 @@
-export * from './ResponsiveChartContainerPro';
