+++ packages/x-charts-pro/src/context/ChartDataProviderPro/index.ts
@@ -1,0 +1,1 @@
+export * from './ChartDataProviderPro';
