+++ packages/x-charts-pro/src/index.ts
@@ -28,7 +28,6 @@ export * from '@mui/x-charts/ChartsSurface';
 
 // Pro components
 export * from './Heatmap';
-export * from './ResponsiveChartContainerPro';
 export * from './ChartContainerPro';
 export * from './ScatterChartPro';
 export * from './BarChartPro';
