+++ packages/x-charts/src/internals/index.ts
@@ -35,6 +35,7 @@ export * from '../context/PluginProvider';
 export * from '../context/AnimationProvider';
 export type * from '../context/context.types';
 export { getAxisExtremum } from '../context/CartesianProvider/getAxisExtremum';
+export * from '../context/ChartDataProvider';
 
 // series configuration
 export * from '../models/seriesType/config';
